+++ src/main/java/com/final_10aeat/domain/admin/dto/request/EmailVerificationRequestDto.java
@@ -1,4 +1,4 @@
-package com.final_10aeat.domain.member.dto.request;
+package com.final_10aeat.domain.admin.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
