+++ src/main/java/com/final_10aeat/domain/admin/service/DevelopEmailService.java
@@ -1,12 +1,11 @@
-package com.final_10aeat.domain.member.service;
+package com.final_10aeat.domain.admin.service;
 
-import com.final_10aeat.domain.member.dto.response.EmailVerificationResponseDto;
+import com.final_10aeat.domain.admin.dto.response.EmailVerificationResponseDto;
 import com.final_10aeat.domain.member.entity.MemberRole;
-import com.final_10aeat.domain.member.exception.EmailSendingException;
-import com.final_10aeat.domain.member.exception.EmailTemplateLoadException;
-import com.final_10aeat.domain.member.exception.InvalidVerificationCodeException;
-import com.final_10aeat.domain.member.exception.VerificationCodeExpiredException;
-import com.final_10aeat.global.util.ResponseDTO;
+import com.final_10aeat.domain.admin.exception.EmailSendingException;
+import com.final_10aeat.domain.admin.exception.EmailTemplateLoadException;
+import com.final_10aeat.domain.admin.exception.InvalidVerificationCodeException;
+import com.final_10aeat.domain.admin.exception.VerificationCodeExpiredException;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import jakarta.mail.MessagingException;
@@ -24,7 +23,6 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.http.HttpStatus;
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
